wip: Split tests into modules by scope
