+++ tests/migration.rs
@@ -1,36 +1,7 @@
-use near_units::parse_near;
-use near_workspaces::AccountId;
+mod test_env;
+
+use crate::test_env::*;
 use serde_json::json;
-
-const DEVHUB_CONTRACT: &str = "devgovgigs.near";
-const NEAR_SOCIAL: &str = "social.near";
-
-async fn init_contracts() -> anyhow::Result<near_workspaces::Contract> {
-    let worker = near_workspaces::sandbox().await?;
-    let mainnet = near_workspaces::mainnet_archival().await?;
-
-    // NEAR social deployment
-    let near_social_id: AccountId = NEAR_SOCIAL.parse()?;
-    let near_social = worker
-        .import_contract(&near_social_id, &mainnet)
-        .initial_balance(parse_near!("10000 N"))
-        .transact()
-        .await?;
-    near_social.call("new").transact().await?.into_result()?;
-
-    // Devhub contract deployment
-    let contract_id: AccountId = DEVHUB_CONTRACT.parse()?;
-    let contract = worker
-        .import_contract(&contract_id, &mainnet)
-        .initial_balance(parse_near!("1000 N"))
-        .transact()
-        .await?;
-    let outcome = contract.call("new").args_json(json!({})).transact().await?;
-    assert!(outcome.is_success());
-    assert!(format!("{:?}", outcome).contains("Migrated to version:"));
-
-    Ok(contract)
-}
 
 #[tokio::test]
 async fn test_deploy_contract_self_upgrade() -> anyhow::Result<()> {
@@ -236,60 +207,6 @@ async fn test_deploy_contract_self_upgrade() -> anyhow::Result<()> {
         .await?
         .json()?;
     insta::assert_json_snapshot!(get_community);
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_posting_features() -> anyhow::Result<()> {
-    let deposit_amount = near_units::parse_near!("0.1");
-
-    let contract = init_contracts().await?;
-
-    // Call self upgrade with current branch code
-    // compile the current code
-    let wasm = near_workspaces::compile_project("./").await?;
-
-    let mut contract_upgrade_result =
-        contract.call("unsafe_self_upgrade").args(wasm).max_gas().transact().await?;
-
-    while contract_upgrade_result.json::<String>()? == "needs-migration" {
-        contract_upgrade_result =
-            contract.call("unsafe_migrate").args_json(json!({})).max_gas().transact().await?;
-    }
-
-    let add_solution_post = contract
-        .call("add_post")
-        .args_json(json!({
-                        "parent_id": null,
-                        "labels": [],
-                        "body": {
-                                        "name": "Solution Test",
-                                        "description": "This is a test solution post.",
-                                        "post_type": "Solution",
-                                        "requested_sponsor": "neardevgov.near",
-                                        "requested_sponsorship_amount": "1000",
-                                        "sponsorship_token": "NEAR",
-                                        "solution_version": "V2"
-                        }
-        }))
-        .deposit(deposit_amount)
-        .max_gas()
-        .transact()
-        .await?;
-
-    assert!(add_solution_post.is_success());
-
-    let get_solution_post: serde_json::Value = contract
-        .call("get_post")
-        .args_json(json!({
-                        "post_id" : 1
-        }))
-        .view()
-        .await?
-        .json()?;
-
-    insta::assert_json_snapshot!(get_solution_post, {".snapshot.timestamp" => "[timestamp]"});
 
     Ok(())
 }
